\name{SeqVarTools-package}
\alias{SeqVarTools-package}
\alias{SeqVarTools}
\docType{package}
\title{Tools for Variant Analysis}

\description{
  This package provides tools for data exploration and analysis of
  variants, extending the functionality of the package
  \pkg{\link{SeqArray}}.
}

\details{
  \pkg{\link{SeqArray}} provides an alternative to the Variant Call
  Format (VCF) for storage of variants called from sequencing data,
  enabling efficient storage, fast access to subsets of the data, and
  rapid computation.

  SeqVarTools provides an interface to the \pkg{\link{SeqArray}} storage
  format with tools for many common tasks in variant analysis and
  integration with basic S4 classes in Bioconductor.
}

\author{
  Stephanie M. Gogarten, Xiuwen Zheng
  
  Maintainer: Stephanie M. Gogarten \email{sdmorris@u.washington.edu}
}

\keyword{package}
